ENH: replaced faith_pd with faith_pd_fast
