+++ skbio/diversity/alpha/__init__.py
@@ -175,7 +175,6 @@ from ._base import (
     singles, strong)
 from ._gini import gini_index
 from ._lladser import lladser_pe, lladser_ci
-from ._fast_faith import faith_pd_fast
 
 __all__ = ['ace', 'chao1', 'chao1_ci', 'berger_parker_d', 'brillouin_d',
            'dominance', 'doubles', 'enspie', 'esty_ci',
@@ -183,6 +182,6 @@ __all__ = ['ace', 'chao1', 'chao1_ci', 'berger_parker_d', 'brillouin_d',
            'kempton_taylor_q', 'margalef', 'mcintosh_d', 'mcintosh_e',
            'menhinick', 'michaelis_menten_fit', 'observed_otus', 'osd',
            'pielou_e', 'robbins', 'shannon', 'simpson', 'simpson_e', 'singles',
-           'strong', 'gini_index', 'lladser_pe', 'lladser_ci', 'faith_pd_fast']
+           'strong', 'gini_index', 'lladser_pe', 'lladser_ci']
 
 test = TestRunner(__file__).test
