+++ skbio/diversity/tests/test_base.py
@@ -16,7 +16,7 @@ import numpy.testing as npt
 
 from skbio.diversity._base import (
     _validate_counts_vector, _validate_counts_vectors,
-    _validate_otu_ids_and_tree)
+    _validate_otu_ids_and_tree, _index_tree, _counts_and_index)
 from skbio import TreeNode
 from skbio.tree import DuplicateNodeError, MissingNodeError
 
@@ -228,6 +228,27 @@ class BaseTests(TestCase):
         self.assertRaises(MissingNodeError, _validate_otu_ids_and_tree, counts,
                           otu_ids, t)
 
+    def test_index_tree(self):
+        t = TreeNode.read(StringIO(u"((a:1, b:2)c:3)root;"))
+        indexed = _index_tree(t)
+        npt.assert_equal(indexed['length'], np.array([1, 2, 3, 0], dtype=float))
+
+    def test_counts_and_index(self):
+        t = TreeNode.read(StringIO(u"((a:1, b:2)c:3)root;"))
+        counts = np.array([[0, 1], [1, 5], [10, 1]])
+        count_array, indexed = _counts_and_index(counts, np.array(['a', 'b']),
+                                                 t, None)
+        exp_counts = np.array([[0, 1, 10], [1, 5, 1], [1, 6, 11], [1, 6, 11]])
+        npt.assert_equal(count_array, exp_counts)
+
+    def test_counts_and_length_with_index(self):
+        t = TreeNode.read(StringIO(u"((a:1, b:2)c:3)root;"))
+        counts = np.array([[0, 1], [1, 5], [10, 1]])
+        indexed = _index_tree(t)
+        count_array, indexed = _counts_and_index(counts, np.array(['a', 'b']),
+                                                 t, indexed)
+        exp_counts = np.array([[0, 1, 10], [1, 5, 1], [1, 6, 11], [1, 6, 11]])
+        npt.assert_equal(count_array, exp_counts)
 
 if __name__ == '__main__':
     main()
